requirements.txt
# Flask Web Framework
Flask

# Data Handling and Manipulation
pandas
numpy
openpyxl

# Document Generation
python-docx
python-pptx

# Statistical Analysis & Data Reading
scipy
pyreadstat

# Charting
matplotlib

# JSON Schema Validation
jsonschema

# Executable Builder
pyinstaller